Moved the yaml initialization to root.go
