+++ pkg/shared/config/config.go
@@ -3,9 +3,38 @@ package config
 import (
 	"fmt"
 	"os"
+	"time"
 
 	yaml "gopkg.in/yaml.v2"
 )
+
+type Config struct {
+	Logger     Logger     `yaml:"logger"`
+	HttpClient HttpClient `yaml:"http_client"`
+}
+
+type Logger struct {
+	Level string `yaml:"level"`
+}
+
+type HttpClient struct {
+	Debug            string          `yaml:"debug"`
+	RetryCount       time.Duration   `yaml:"retry_count"`
+	RetryWaitTime    time.Duration   `yaml:"retry_wait_time"`
+	RetryMaxWaitTime time.Duration   `yaml:"retry_max_wait_time"`
+	Timeout          time.Duration   `yaml:"timeout"`
+	TlsClientConfig  TlsClientConfig `yaml:"tls_client_config"`
+	Proxy            Proxy           `yaml:"proxy"`
+}
+
+type TlsClientConfig struct {
+	Verify bool `yaml:"verify"`
+}
+
+type Proxy struct {
+	Host string `yaml:"host"`
+	Port string `yaml:"port"`
+}
 
 func ValidateConfigPath(path string) error {
 	s, err := os.Stat(path)
@@ -13,7 +42,7 @@ func ValidateConfigPath(path string) error {
 		return err
 	}
 	if s.IsDir() {
-		return fmt.Errorf("'%s' is a directory, not a normal file", path)
+		return fmt.Errorf("'%s' is a directory, not a file", path)
 	}
 	return nil
 }
@@ -39,11 +68,6 @@ func LoadYAML(configPath string, data interface{}) error {
 
 func NewConfig(configPath string) (*Config, error) {
 	config := &Config{}
-
-	//TODO: move to default internal configuration
-	// if len(configPath) == 0 {
-
-	// }
 
 	if err := LoadYAML(configPath, &config); err != nil {
 		return nil, err
